s3: rename getStreamObject to getObjectStream
