+++ packages/s3/index.js
@@ -5,4 +5,4 @@ export { default as getObject } from './cmd/getObject.js'
 export { default as putObject } from './cmd/putObject.js'
 
 // Using streams
-export { default as getStreamObject } from './cmd/getStreamObject.js'
+export { default as getObjectStream } from './cmd/getObjectStream.js'
