+++ packages/s3/manifest.ttl
@@ -3,13 +3,6 @@
 @prefix p: <https://pipeline.described.at/> .
 @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
 
-<getObject> a p:Operation;
-  rdfs:label "Get S3 object";
-  rdfs:comment "Get an object from a S3 bucket.";
-  code:implementedBy [ a code:EcmaScriptModule;
-    code:link <node:barnard59-s3/cmd/getObject.js#default>
-  ].
-
 <putObject> a p:Operation;
   rdfs:label "Put an object to S3";
   rdfs:comment "Put an object to a S3 bucket.";
@@ -17,9 +10,16 @@
     code:link <node:barnard59-s3/cmd/putObject.js#default>
   ].
 
-<getStreamObject> a p:Operation, p:Readable;
+<getObject> a p:Operation;
+  rdfs:label "Get S3 object";
+  rdfs:comment "Get an object from a S3 bucket.";
+  code:implementedBy [ a code:EcmaScriptModule;
+    code:link <node:barnard59-s3/cmd/getObject.js#default>
+  ].
+
+<getObject/stream> a p:Operation, p:Readable;
   rdfs:label "Get S3 object as stream";
   rdfs:comment "Get an object from a S3 bucket as a Readable stream.";
   code:implementedBy [ a code:EcmaScriptModule;
-    code:link <node:barnard59-s3/cmd/getStreamObject.js#default>
+    code:link <node:barnard59-s3/cmd/getObjectStream.js#default>
   ].
